backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
PyMySQL==1.1.0
requests==2.31.0
reportlab==4.1.0
python-dotenv==1.0.1
whitenoise==6.6.0
gunicorn==21.2.0
cryptography==42.0.5